requirements.txt
langchain==0.0.265
openai==0.27.8
fastapi==0.101.1
uvicorn==0.23.2
googlesearch-python==1.2.3
duckduckgo-search==3.8.5
python-dotenv==1.0.0
requests~=2.31.0
pydantic~=1.10.12